auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = "Authorization & authentification service"
authors = ["abolychev <bolychev.anton@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.110.0"}
pyjwt = "^2.8.0"
passlib = "^1.7.4"
sqlmodel = "^0.0.16"
common = {path = "../common", develop = true}
nats-py = "^2.7.2"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
